+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/person/ai/mission/ConstructionMission.java
@@ -32,6 +32,7 @@ import org.mars_sim.msp.core.person.Person;
 import org.mars_sim.msp.core.person.ai.SkillType;
 import org.mars_sim.msp.core.person.ai.task.ConstructBuilding;
 import org.mars_sim.msp.core.person.ai.task.DigLocalRegolith;
+import org.mars_sim.msp.core.person.ai.task.EVAOperation;
 import org.mars_sim.msp.core.person.ai.task.util.Task;
 import org.mars_sim.msp.core.person.ai.task.util.Worker;
 import org.mars_sim.msp.core.resource.ItemResourceUtil;
@@ -655,7 +656,7 @@ public class ConstructionMission extends AbstractMission
 			
 			Task currentTask = p.getMind().getTaskManager().getTask();
 			if (currentTask != null && !currentTask.getName().equalsIgnoreCase(DigLocalRegolith.NAME)) {
-				boolean accepted = p.getMind().getTaskManager().addPendingTask(DigLocalRegolith.NAME);
+				boolean accepted = p.getMind().getTaskManager().addPendingTask(DigLocalRegolith.SIMPLE_NAME);
 				if (accepted)
 					logger.info(p, 60_000, "Confirmed receiving the assigned task of DigLocalRegolith.");
 			}